requirements.txt
python-hcl2==4.3.2